requirements.txt
pymodbus
pyyaml
flake8
aiosolaredge
